refactor(ui): extract single participant tile and simplify LiveKit
Replace the large, monolithic livekit-participant-tile implementation with a streamlined version that delegates per-participant rendering/state to livekit-single-participant-tile. Removed many unused imports, verbose controls, and the heavy options panel logic; simplified the Zod schema and props to only the needed fields. This breaks the component into smaller concerns, reduces bundle size, and makes participant-specific behavior easier to maintain and test.
